chore: redistribute registered oracles files
